[profiling] Extract C API module; refactor Uuuid
Extracts the bits for the C API of the profiler into one location.
This API surface will expand with Endpoint Profiling and this is
a cleanup PR in preparation for that.

Also simplify capi::Uuid and add tests.
